refactor(db): extract pool management to src/db/pool.ts
Move getPool and closePool into their own module so pool lifecycle is
isolated from query logic. Both are re-exported from db.ts unchanged.
